refactor: app 레이어 provider, styles 단위 정리
