+++ src/app/_styles/GlobalStyle.ts
@@ -3,7 +3,7 @@
 import { createGlobalStyle } from 'styled-components';
 import reset from 'styled-reset';
 
-const GlobalStyle = createGlobalStyle`
+export const GlobalStyle = createGlobalStyle`
     ${reset}
     * {   
         box-sizing: border-box;
@@ -87,5 +87,3 @@ const GlobalStyle = createGlobalStyle`
     }
     
 `;
-
-export default GlobalStyle;
