+++ src/app/_styles/defaultTheme.ts
@@ -1,0 +1,12 @@
+const defaultTheme = {
+  color: {
+    MAIN_BG: '#1f223f',
+    WHILE: '#FFFFFF',
+    MAIN_COLOR: '#3ec578',
+  },
+  padding: {
+    container: 20,
+  },
+};
+
+export default defaultTheme;
